+++ src/main/java/ru/yandex/practicum/filmorate/exception/DuplicateException.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.exeption;
+package ru.yandex.practicum.filmorate.exception;
 
 public class DuplicateException extends RuntimeException {
