+++ src/main/java/ru/yandex/practicum/filmorate/exception/NotFoundException.java
@@ -1,0 +1,19 @@
+package ru.yandex.practicum.filmorate.exception;
+
+public class NotFoundException extends RuntimeException {
+
+    public NotFoundException() {
+    }
+
+    public NotFoundException(final String message) {
+        super(message);
+    }
+
+    public NotFoundException(final String message, final Throwable cause) {
+        super(message, cause);
+    }
+
+    public NotFoundException(final Throwable cause) {
+        super(cause);
+    }
+}
