+++ src/main/java/ru/yandex/practicum/filmorate/storage/LikeStorage.java
@@ -1,0 +1,17 @@
+package ru.yandex.practicum.filmorate.storage;
+
+import ru.yandex.practicum.filmorate.model.Genre;
+
+import java.util.List;
+
+public interface LikeStorage {
+    void addLike(Long filmId, Long userId);
+
+    void removeLike(Long filmId, Long userId);
+
+    List<Long> getLikes(Long filmId);
+
+    Long getLikesAmount(Long filmId);
+
+    List<Long> getPopularFilmsId(int size);
+}
